\name{mixedlang-package}
\alias{mixedlang-package}
\alias{mixedlang}
\docType{package}
\title{
\packageTitle{mixedlang}
}
\description{
\packageDescription{mixedlang}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mixedlang}
\packageIndices{mixedlang}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mixedlang}

Maintainer: \packageMaintainer{mixedlang}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
