% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_function}
\alias{test_function}
\title{Test of calling a Fortran Function}
\usage{
test_function(x)
}
\arguments{
\item{x}{A numeric vectir}
}
\value{
A transformed numeric vector
}
\description{
Test of calling a Fortran Function
}
\examples{
test_function(0:9)
}
